requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
psycopg2-binary>=2.9.9
neo4j>=5.14.0
sqlalchemy>=2.0.23

# NLP and ML
spacy>=3.7.0
nltk>=3.8.1
transformers>=4.35.0
torch>=2.1.0
sentencepiece>=0.1.99

# Computer Vision (optional, for VVA)
opencv-python>=4.8.1
Pillow>=10.1.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
redis>=5.0.1
python-multipart>=0.0.6
tweepy>=4.14.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
